fix bugs

+++ database/migrations/2024_05_30_004701_create_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price', 8, 2);
             $table->text('description');
             $table->text('img');
             $table->timestamps();
