+++ database/migrations/2024_05_30_005754_create_guests_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->date('birthData');
-            $table->smallInteger('identificationNumber');
+            $table->integer('identificationNumber');
             $table->timestamps();
         });
     }
